useBeforeScreenshot
