+++ app/packages/looker-3d/package.json
@@ -26,6 +26,7 @@
         "@types/node": "^18.0.3",
         "@types/react": "^18.0.0",
         "@types/react-dom": "^18.0.0",
+        "@types/three": "^0.146.0",
         "@vitejs/plugin-react": "^1.3.0",
         "nodemon": "^2.0.19",
         "rollup-plugin-external-globals": "^0.6.1",
